+++ .babelrc
@@ -1,0 +1,10 @@
+{
+    "presets": [
+        "@babel/preset-env",
+        ["@babel/preset-react", {"runtime": "automatic"}]
+    ],
+    "plugins":[
+        ["@babel/transform-runtime"],
+        ["import", { "libraryName": "antd", "libraryDirectory": "lib", "style": "css" }]
+    ]
+  }
