+++ .production.env
@@ -1,0 +1,3 @@
+HOST_SHIPMENT=http://localhost:5020
+HOST_UIPARS=http://localhost:5050
+TOKEN=eyJhbGciOiJodHRwOi8vd3d3LnczLm9yZy8yMDAxLzA0L3htbGRzaWctbW9yZSNobWFjLXNoYTI1NiIsInR5cCI6IkpXVCJ9.eyJBY2NvdW50IjoidGVzdDEiLCJJRCI6IjAyMDBhYTJmLTE1ZDUtNGMzMS05NmQ0LTU0ZTVlYmFjNzI2MSIsIklEUGFydG5lciI6IlhMT0ciLCJDb2RlV2FyZWhvdXNlIjoiV0gwMSIsImV4cCI6MTY1ODIyNDkzOSwiaXNzIjoiaHR0cDovL2xvY2FsaG9zdDo1MDUwIiwiYXVkIjoiaHR0cDovL2xvY2FsaG9zdDo1MDUwIn0.m4ovNilqeQL6FKMCVsDebMUnQ3Q8Gk3gVBbyYYT0bRM
