+++ package.json
@@ -19,7 +19,11 @@
     "start": "react-scripts start",
     "build": "react-scripts build",
     "test": "react-scripts test",
-    "eject": "react-scripts eject"
+    "eject": "react-scripts eject",
+    "dev": "set NODE_ENV=development && webpack-dev-server --mode development --open --hot",
+    "build-dev": "set NODE_ENV=development && webpack --mode development",
+    "product": "set NODE_ENV=production && webpack-dev-server --mode production --open --hot",
+    "build-product": "set NODE_ENV=production && webpack --mode production"
   },
   "eslintConfig": {
     "extends": [
@@ -38,5 +42,21 @@
       "last 1 firefox version",
       "last 1 safari version"
     ]
+  },
+  "devDependencies": {
+    "babel-loader": "^8.2.5",
+    "babel-plugin-import": "^1.13.5",
+    "css-loader": "^6.7.1",
+    "dotenv-webpack": "^7.1.0",
+    "html-loader": "^3.1.0",
+    "html-webpack-plugin": "^5.5.0",
+    "mini-css-extract-plugin": "^2.6.0",
+    "sass": "^1.52.2",
+    "sass-loader": "^13.0.0",
+    "webpack": "^5.73.0",
+    "webpack-ant-icon-loader": "^1.0.8",
+    "webpack-bundle-analyzer": "^4.5.0",
+    "webpack-cli": "^4.9.2",
+    "webpack-dev-server": "^4.9.2"
   }
 }
