+++ src/index.js
@@ -4,7 +4,6 @@ import 'antd/dist/antd.css';
 import './index.css';
 import ManageWarehouseImportExportOrder from './App';
 //import reportWebVitals from './reportWebVitals';
-
 const root = ReactDOM.createRoot(document.getElementById('root'));
 root.render(
     <ManageWarehouseImportExportOrder />
